Replace TypeManagement with dTypeBrowse, using vuex store
