+++ client/src/views/dTypeBrowse.vue
@@ -1,65 +1,335 @@
 <template>
-    <v-container>
-        <TypeManagement
-            :contract="contract"
-            :from="from"
-        />
-    </v-container>
+    <div>
+        <v-toolbar flat color='white'>
+            <v-toolbar-title>dTypes</v-toolbar-title>
+            <v-spacer></v-spacer>
+            <v-dialog v-model='dialog' max-width='600px'>
+                <template v-slot:activator='{ on }'>
+                    <v-btn
+                        small flat fab
+                        v-on='on'
+                    >
+                        <v-icon>fa-plus</v-icon>
+                    </v-btn>
+                </template>
+                <v-card>
+                    <v-card-title>
+                        <span class='headline'>{{ formTitle }}</span>
+                    </v-card-title>
+
+                    <v-card-text v-if="editedItem">
+                        <v-container grid-list-md>
+                            <v-layout wrap>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.name' label='name'></v-text-field>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-layout row wrap>
+                                    <v-flex xs6>
+                                    <!-- <v-text-field v-model='editedItem.types' label='types'></v-text-field> -->
+                                    <v-chip close
+                                        v-for="(type, i) in editedItem.types"
+                                        v-on:input="onRemovedType(type, i)"
+                                    >
+                                        {{type}}
+                                    </v-chip>
+                                    </v-flex>
+                                    <v-flex xs6>
+                                    <dTypeSearch
+                                        label='types'
+                                        itemKey='name'
+                                        itemValue='name'
+                                        :items='dtypes'
+                                        v-on:change="onSelectedType"
+                                    />
+                                    </v-flex>
+                                </v-layout>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.labels' label='labels'></v-text-field>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.lang' label='language'></v-text-field>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.isEvent' label='isEvent'></v-text-field>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.isFunction' label='isFunction'></v-text-field>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.hasOutput' label='hasOutput'></v-text-field>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.contractAddress' label='address'></v-text-field>
+                                </v-flex>
+                                <v-flex xs12>
+                                    <v-text-field v-model='editedItem.source' label='source'></v-text-field>
+                                </v-flex>
+                            </v-layout>
+                        </v-container>
+                    </v-card-text>
+
+                    <v-card-actions>
+                        <v-spacer></v-spacer>
+                        <v-btn color='blue darken-1' flat @click='close'>Cancel</v-btn>
+                        <v-btn color='blue darken-1' flat @click='save'>Save</v-btn>
+                    </v-card-actions>
+                </v-card>
+            </v-dialog>
+            <v-dialog v-model='dialogBulk' max-width='600px'>
+                <template v-slot:activator='{ on }'>
+                    <v-btn
+                        small flat fab
+                        v-on='on'
+                    >
+                        <v-icon>fa-folder-plus</v-icon>
+                    </v-btn>
+                </template>
+                <v-card>
+                    <v-card-title>
+                        <span class='headline'>Bulk Insert</span>
+                    </v-card-title>
+
+                    <v-card-text>
+                        <v-container grid-list-md>
+                            <v-layout wrap>
+                                <v-flex xs12>
+                                    <v-textarea
+                                        outline
+                                        name="bulkInsert"
+                                        v-model='bulkInsert'
+                                        label="array of types"
+                                        :placeholder="bulkInsertDefault"
+                                    ></v-textarea>
+                                </v-flex>
+                            </v-layout>
+                        </v-container>
+                    </v-card-text>
+
+                    <v-card-actions>
+                        <v-spacer></v-spacer>
+                        <v-btn color='blue darken-1' flat @click='closeBulk'>Cancel</v-btn>
+                        <v-btn color='blue darken-1' flat @click='saveBulk'>Save</v-btn>
+                    </v-card-actions>
+                </v-card>
+            </v-dialog>
+        </v-toolbar>
+        <v-data-table
+            :headers='headers'
+            :items='dtypes'
+            class='elevation-1'
+        >
+            <template v-slot:items='props'>
+
+                <template
+                    v-for="header in headers"
+                >
+                    <td class='text-xs-left'>
+                        <dTypeBrowseField
+                            :field="header.value"
+                            :value="props.item[header.value]"
+                        />
+                    </td>
+                </template>
+                <td class='justify-center layout px-0'>
+                    <v-icon
+                        small
+                        class='mr-2'
+                        @click='goRoute(props.item.typeHash)'
+                    >
+                        launch
+                    </v-icon>
+                    <v-icon
+                        small
+                        class='mr-2'
+                        @click='editItem(props.item)'
+                    >
+                        edit
+                    </v-icon>
+                    <v-icon
+                        small
+                        @click='deleteItem(props.item)'
+                    >
+                        delete
+                    </v-icon>
+                </td>
+            </template>
+        </v-data-table>
+    </div>
 </template>
 
 <script>
-import {ethers} from 'ethers';
-import TypeManagement from '../components/TypeManagement';
-import DType from '../constants';
-import {waitAsync} from '../utils';
+import dTypeSearch from '../components/dTypeSearch';
+import dTypeBrowseField from '../components/dTypeBrowseField';
 
 export default {
     name: 'dTypeBrowse',
     components: {
-        TypeManagement,
+        dTypeSearch,
+        dTypeBrowseField,
     },
-    data() {
-        return {
-            contract: null,
-            from: DType.from,
-        };
+    data: () => ({
+        dialog: false,
+        dialogBulk: false,
+        editedIndex: -1,
+        editedItem: {
+            name: '',
+            types: [],
+            labels: [],
+            lang: 0,
+            isEvent: false,
+            isFunction: false,
+            hasOutput: false,
+            contractAddress: '0x0000000000000000000000000000000000000000',
+            source: '0x0000000000000000000000000000000000000000000000000000000000000000',
+        },
+        defaultItem: {
+            name: '',
+            types: [],
+            labels: [],
+            lang: 0,
+            isEvent: false,
+            isFunction: false,
+            hasOutput: false,
+            contractAddress: '0x0000000000000000000000000000000000000000',
+            source: '0x0000000000000000000000000000000000000000000000000000000000000000',
+        },
+        bulkInsert: JSON.stringify([{
+                name: "uint256",
+                types: [],
+                labels: [],
+                lang: 0,
+                isEvent: false,
+                isFunction: false,
+                hasOutput: false,
+                contractAddress: "0xCd9492Cdae7E8F8B5a648c6E15c4005C4cd9028b",
+                source: "0x0000000000000000000000000000000000000000000000000000000000000000"
+        }]),
+        bulkInsertDefault: JSON.stringify([{
+                name: "uint256",
+                types: [],
+                labels: [],
+                lang: 0,
+                isEvent: false,
+                isFunction: false,
+                hasOutput: false,
+                contractAddress: "0xCd9492Cdae7E8F8B5a648c6E15c4005C4cd9028b",
+                source: "0x0000000000000000000000000000000000000000000000000000000000000000"
+        }]),
+    }),
+    computed: {
+        contract() {
+            console.log('dTypeBrowse contract', this.$store.state.contract);
+            return this.$store.state.contract;
+        },
+        dtype() {
+            console.log('dTypeBrowse dtype', this.$store.state.dtype);
+            return this.$store.state.dtype;
+        },
+        dtypes() {
+            return this.$store.state.dtypes;
+        },
+        headers() {
+            return !this.dtype ? [] : this.dtype.data.labels.map(
+                label => ({text: label, value: label})
+            );
+        },
+        formTitle() {
+            return this.editedIndex === -1 ? 'New Item' : 'Edit Item';
+        },
     },
-    created() {
-        this.setStage();
+    watch: {
+        dialog (val) {
+            val || this.close();
+        },
+        dialogBulk (val) {
+            val || this.closeBulk();
+        },
+        dtypes() {
+            this.setBulkInsert();
+        },
     },
     methods: {
-        async setStage() {
-            // const provider = new ethers.providers.JsonRpcProvider('http://127.0.0.1:9545');
-
-            // Metamask
-            let provider = new ethers.providers.Web3Provider(web3.currentProvider);
-            let wallet = new ethers.Wallet(DType.from.privateKey, provider);
-
-            await waitAsync(1000);
-
-            // Temporary fix for Ganache
-            if (provider.network.chainId == 5777) {
-                const url = "http://localhost:8545";
-                provider = new ethers.providers.JsonRpcProvider(url);
-                wallet = provider.getSigner(0);
-                await waitAsync(1000);
-            }
-
-            console.log(provider);
-            console.log(wallet);
-            console.log('DType', DType);
-
-            const contractAddress = DType.contract.networks[
-                String(provider.network.chainId)
-            ].address;
-
-            console.log('contractAddress', contractAddress);
-            this.contract = new ethers.Contract(
-                contractAddress,
-                DType.contract.abi,
-                wallet,
+        setBulkInsert() {
+            this.bulkInsert = JSON.stringify(
+                this.dtypes.map(dt => {
+                    let obj = {};
+                    Object.keys(this.defaultItem).forEach(key => {
+                        obj[key] = dt[key];
+                    });
+                    obj.typeHash = dt.typeHash;
+                    return obj;
+                })
             );
+        },
+        async insert(dtype) {
+            this.$store.dispatch('insertType', dtype);
+        },
+        async batchInsert(dtypeArray) {
+            this.$store.dispatch('insertBatchType', dtypeArray);
+        },
+        async update(dtype) {
+            this.$store.dispatch('updateType', dtype);
+        },
+        async delete(dtype) {
+            this.$store.dispatch('removeType', dtype);
+        },
+        editItem(item) {
+            this.editedIndex = this.dtypes.indexOf(item)
+            this.editedItem = Object.assign({}, item)
+            this.dialog = true
+        },
+        deleteItem(item) {
+            const index = this.dtypes.indexOf(item)
+            confirm('Are you sure you want to delete this item?') && this.delete(item);
+        },
+        close() {
+            this.dialog = false;
+            setTimeout(() => {
+                this.editedItem = Object.assign({}, this.defaultItem);
+                this.editedIndex = -1;
+            }, 300)
+        },
+        save() {
+            this.editedItem.labels = this.editedItem.labels.split(',');
+            if (this.editedIndex > -1) {
+                this.update(this.editedItem);
+            } else {
+                this.insert(this.editedItem);
+            }
+            this.close();
+        },
+        closeBulk() {
+            this.dialogBulk = false;
+        },
+        saveBulk() {
+            let bulk;
+            try {
+                bulk = JSON.parse(this.bulkInsert);
+                this.batchInsert(bulk);
+                this.closeBulk();
+            } catch (e) {
+                alert(`${e}`);
+            }
+        },
+        onSelectedType(values) {
+            values = this.editedItem.types.concat(values);
+            this.editedItem.types = values;
+        },
+        onRemovedType(value, i) {
+            this.editedItem.types.splice(i, 1);
+        },
+        goRoute(typeHash) {
+            let route = this.$router.resolve({path: `dtype/${typeHash}`});
+            window.open(route.href, '_blank');
         },
     },
 };
 </script>
+
+<style>
+.container {
+    max-width: 100%;
+}
+</style>
